Refactor fetchIpAdressFromRequest into public utils func
